Refactor internal structures for the compared JSON
The existing structures had some duplications and limitations.
In the old model, if it was not necessary an array or object was
not transformed to our internal structures, thus it was left as
a raw JSON. As consequence of this, it was hard to reuse existing
UI elements to make this raw JSON value collapsable. In the new
model, we try to reuse functionality and distinguish different
internal types making them easier to be worked with.

This commit fixes #2, i.e. making all arrays and objects collapsable
in the UI.